Prevent user from exceeding 'maxlength'
The 'type' function does now check for the input's 'maxlength' attribute
to prevent the user from typing more than a certain number of characters
by using the plugin's keyboard.

If 'maxlength' attribute is available, when the user tries to append a new
character by clicking on a key, the input will not be updated unless its
current value's length is lower than 'maxlength'.

Behaviour when the user has higlighted part of the input's content won't
be affected by this change, as replacing existing contents won't cause the
value's length to grow.