Modified Backspace function
Fixed the backspace function to make it function normally. Previously the back button moved the cursor to the beginning of the line, and/or always removed the character from the end of the line rather than the character in front of the cursor. This was referenced in issue #7 of javidan/jkeyboard